azure-data-pipeline/api/dashboard-api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
azure-cosmos==4.5.1
paho-mqtt==1.6.1
python-dotenv==1.0.0
